intra-django/requirements.txt
asgiref==3.4.1
Django==3.2.12
djangorestframework==3.13.1
gunicorn==20.1.0
mysqlclient==2.1.0
pkg-resources==0.0.0
pytz==2021.3
sqlparse==0.4.2
typing-extensions==4.1.1
